Delete MainContent.
