+++ src/pages/index.tsx
@@ -1,9 +1,12 @@
 import { NextPage } from "next";
 import HeaderBar from "@/components/molecules/HeaderBar";
-import MainContent from "@/components/organisms/MainContent";
 import FooterNavigation from "@/components/organisms/FooterNavigation";
 import { makeStyles } from "@material-ui/styles";
 import { Theme } from "@material-ui/core/styles/createMuiTheme";
+import HeaderTab from "@/components/molecules/HeaderTab";
+import TabContents from "@/components/organisms/TabContents";
+import AnimeList from "@/components/organisms/AnimeList";
+import React from "react";
 
 const useStyles = makeStyles((theme: Theme) => ({
   root: {
@@ -18,11 +21,26 @@ const useStyles = makeStyles((theme: Theme) => ({
 }));
 
 const Index: NextPage = () => {
+  const [value, setValue] = React.useState(0);
+
+  function handleTapped(event: React.ChangeEvent<{}>, newValue: number): void {
+    setValue(newValue);
+  }
+
+  function handleSwiped(index: number): void {
+    setValue(index);
+  }
+
+  const weekend = ["月", "火", "水", "木", "金", "土", "日"];
   const classes = useStyles();
+
   return (
     <div className={classes.root}>
       <HeaderBar />
-      <MainContent />
+      <HeaderTab label={weekend} value={value} onTapped={handleTapped} />
+      <TabContents value={value} onSwiped={handleSwiped} count={weekend.length}>
+        <AnimeList times={8} />
+      </TabContents>
       <FooterNavigation />
     </div>
   );
